delivery/requirements.txt
django
djangorestframework
pillow
django-cors-headers
unidecode
drf-yasg
django-filter
djangorestframework-simplejwt
